adk_training/module_12_router_agent/requirements.txt
# Module 12: Router Agent Dependencies
google-adk>=1.0.0
google-cloud-aiplatform>=1.50.0
python-dotenv>=1.0.0

